refactor(tracking): update inputNumberConfig utility import
